Make getMaxSpendable reusable
